requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
ImageIO==2.37.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
scikit-image==0.26.0
scipy==1.17.0
six==1.17.0
tifffile==2025.12.20
